+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Bump iOS deployment target to iOS 9 to match Eigen and avoid adding more deprecated method usage - alloy
 - Adds saved shows page - Kieran
 - Make pre-heated GraphQL cache work for people that upgrade the app - alloy
+- Make content size of event list known upfront to prevent scroller from jumping - alloy
 
 ### 1.8.14 - 1.8.19